Move common cross-sell sheet UI to cross-sell module
