+++ app/feature/feature-home/src/main/kotlin/com/hedvig/android/feature/home/home/ui/HomeDestination.kt
@@ -5,7 +5,6 @@ import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.core.updateTransition
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.MutableWindowInsets
 import androidx.compose.foundation.layout.PaddingValues
@@ -22,7 +21,6 @@ import androidx.compose.foundation.layout.only
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeDrawing
 import androidx.compose.foundation.layout.systemBars
-import androidx.compose.foundation.layout.windowInsetsBottomHeight
 import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
@@ -57,22 +55,19 @@ import androidx.compose.ui.unit.sp
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.compose.dropUnlessResumed
 import androidx.lifecycle.withStateAtLeast
 import arrow.core.nonEmptyListOf
 import com.google.accompanist.permissions.isGranted
 import com.hedvig.android.compose.pager.indicator.HorizontalPagerIndicator
 import com.hedvig.android.compose.ui.preview.BooleanCollectionPreviewParameterProvider
 import com.hedvig.android.core.common.safeCast
-import com.hedvig.android.crosssells.CrossSellsSection
+import com.hedvig.android.crosssells.CrossSellSheet
+import com.hedvig.android.crosssells.CrossSellSheetData
 import com.hedvig.android.data.addons.data.TravelAddonBannerInfo
 import com.hedvig.android.data.addons.data.TravelAddonBannerSource
 import com.hedvig.android.data.contract.android.CrossSell
 import com.hedvig.android.data.contract.android.CrossSell.CrossSellType.ACCIDENT
-import com.hedvig.android.design.system.hedvig.ButtonDefaults
 import com.hedvig.android.design.system.hedvig.ButtonDefaults.ButtonStyle.Secondary
-import com.hedvig.android.design.system.hedvig.FeatureAddonBanner
-import com.hedvig.android.design.system.hedvig.HedvigBottomSheet
 import com.hedvig.android.design.system.hedvig.HedvigBottomSheetState
 import com.hedvig.android.design.system.hedvig.HedvigButton
 import com.hedvig.android.design.system.hedvig.HedvigErrorSection
@@ -132,7 +127,6 @@ import kotlinx.coroutines.flow.drop
 import kotlinx.coroutines.withContext
 import kotlinx.datetime.Instant
 import kotlinx.datetime.toJavaLocalDate
-import octopus.type.UpsellTravelAddonFlow
 
 @Composable
 internal fun HomeDestination(
@@ -200,7 +194,7 @@ private fun HomeScreen(
     refreshingOffset = PullRefreshDefaults.RefreshingOffset + systemBarInsetTopDp,
   )
   val travelBannerInfo = uiState.safeCast<Success>()?.travelAddonBannerInfo
-  val crossSellBottomSheetState = rememberHedvigBottomSheetState<List<CrossSell>>()
+  val crossSellBottomSheetState = rememberHedvigBottomSheetState<CrossSellSheetData>()
   CrossSellBottomSheet(
     state = crossSellBottomSheetState,
     forceShowCrossSells = uiState.forceShowCrossSells,
@@ -271,7 +265,9 @@ private fun HomeScreen(
               )
 
               is HomeTopBarAction.CrossSellsAction -> ToolbarCrossSellsIcon(
-                onClick = { crossSellBottomSheetState.show(action.crossSells) },
+                onClick = {
+                  crossSellBottomSheetState.show(CrossSellSheetData(action.crossSells, uiState.travelAddonBannerInfo))
+                },
               )
 
               is HomeTopBarAction.FirstVetAction -> {
@@ -589,7 +585,7 @@ private fun WelcomeMessage(homeText: HomeText, modifier: Modifier = Modifier) {
 
 @Composable
 private fun CrossSellBottomSheet(
-  state: HedvigBottomSheetState<List<CrossSell>>,
+  state: HedvigBottomSheetState<CrossSellSheetData>,
   travelAddonBannerInfo: TravelAddonBannerInfo?,
   onNavigateToAddonPurchaseFlow: (List<String>) -> Unit,
   forceShowCrossSells: List<CrossSell>?,
@@ -600,7 +596,7 @@ private fun CrossSellBottomSheet(
   LaunchedEffect(forceShowCrossSells, lifecycle) {
     lifecycle.withStateAtLeast(Lifecycle.State.RESUMED) {
       if (forceShowCrossSells?.isNotEmpty() == true) {
-        state.show(forceShowCrossSells)
+        state.show(CrossSellSheetData(forceShowCrossSells, travelAddonBannerInfo))
       }
     }
   }
@@ -613,73 +609,10 @@ private fun CrossSellBottomSheet(
         }
       }
   }
-  HedvigBottomSheet(
-    hedvigBottomSheetState = state,
-    content = { crossSells ->
-      CrossSellsSheetContent(
-        crossSells = crossSells,
-        travelAddonBannerInfo = travelAddonBannerInfo,
-        onCrossSellClick = onCrossSellClick,
-        onNavigateToAddonPurchaseFlow = onNavigateToAddonPurchaseFlow,
-        dismissSheet = { state.dismiss() },
-      )
-    },
-  )
-}
-
-@Suppress("UnusedReceiverParameter")
-@Composable
-private fun ColumnScope.CrossSellsSheetContent(
-  crossSells: List<CrossSell>,
-  travelAddonBannerInfo: TravelAddonBannerInfo?,
-  onCrossSellClick: (String) -> Unit,
-  onNavigateToAddonPurchaseFlow: (List<String>) -> Unit,
-  dismissSheet: () -> Unit,
-) {
-  HedvigText(stringResource(R.string.CROSS_SELL_TITLE))
-  HedvigText(stringResource(R.string.CROSS_SELL_SUBTITLE), color = HedvigTheme.colorScheme.textSecondary)
-  Spacer(Modifier.height(24.dp))
-  CrossSellsSection(
-    showNotificationBadge = false,
-    crossSells = crossSells,
+  CrossSellSheet(
+    state = state,
     onCrossSellClick = onCrossSellClick,
-    withSubHeader = false,
-  )
-  if (travelAddonBannerInfo != null) {
-    Spacer(Modifier.height(24.dp))
-    TravelAddonBanner(
-      travelAddonBannerInfo = travelAddonBannerInfo,
-      launchAddonPurchaseFlow = { eligibleInsurancesIds ->
-        dismissSheet()
-        onNavigateToAddonPurchaseFlow(eligibleInsurancesIds)
-      },
-      modifier = Modifier.fillMaxWidth(),
-    )
-  }
-  Spacer(Modifier.height(24.dp))
-  HedvigButton(
-    text = stringResource(R.string.general_close_button),
-    onClick = dismissSheet,
-    enabled = true,
-    buttonStyle = ButtonDefaults.ButtonStyle.Ghost,
-    modifier = Modifier.fillMaxSize(),
-  )
-  Spacer(Modifier.windowInsetsBottomHeight(WindowInsets.safeDrawing))
-}
-
-@Composable
-private fun TravelAddonBanner(
-  travelAddonBannerInfo: TravelAddonBannerInfo,
-  launchAddonPurchaseFlow: (ids: List<String>) -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  FeatureAddonBanner(
-    modifier = modifier,
-    title = travelAddonBannerInfo.title,
-    description = travelAddonBannerInfo.description,
-    buttonText = stringResource(R.string.ADDON_FLOW_SEE_PRICE_BUTTON),
-    labels = travelAddonBannerInfo.labels,
-    onButtonClick = dropUnlessResumed { launchAddonPurchaseFlow(travelAddonBannerInfo.eligibleInsurancesIds) },
+    onNavigateToAddonPurchaseFlow = onNavigateToAddonPurchaseFlow,
   )
 }
 
@@ -777,30 +710,6 @@ private fun PreviewHomeScreen(
         markCrossSellsNotificationAsSeen = {},
         onNavigateToAddonPurchaseFlow = {},
       )
-    }
-  }
-}
-
-@HedvigPreview
-@Composable
-private fun PreviewCrossSellsSheetContent() {
-  HedvigTheme {
-    Surface(color = HedvigTheme.colorScheme.backgroundPrimary) {
-      Column {
-        CrossSellsSheetContent(
-          crossSells = listOf(CrossSell("rf", "erf", "", "", ACCIDENT)),
-          travelAddonBannerInfo = TravelAddonBannerInfo(
-            title = "Title",
-            description = "description",
-            labels = listOf("Label"),
-            eligibleInsurancesIds = nonEmptyListOf("id"),
-            bannerSource = UpsellTravelAddonFlow.APP_UPSELL_UPGRADE,
-          ),
-          onCrossSellClick = {},
-          onNavigateToAddonPurchaseFlow = {},
-          dismissSheet = {},
-        )
-      }
     }
   }
 }
